BoatData: 1) fixed issue where d_filter_threshold and w_filter_threshold was getting set to masked GPSData: 1) forced diff quality, hdop, and num_sats to be float Measurement: 1) fixed issue with xml output for a user supplied constant temperature Python2Matlab: 1) added code to catch additional None values QComp: 1) fixed issue selecting correct loop test for discharge correction TransectData: 1) added code to handle situation reading raw GPS data that have only 1 value per ensemble, this was for Nortek data WaterData: 1) fixed sl_cutoff based on vertical beam equation. Add (), 2) fixed issue where d_filter_threshold and w_filter_threshold was getting set to masked, 3) fixed issue with invalid shiptrack at beginning or end of transect. BoatSpeed: 1) removed invalid_bt from parameters and use of the parameter in the code, 2) changed plotting of invalid data symbols to reflect location of invalid data, 3) fixed issue turning on/off GPS invalid data symbols GPSFilters: 1) fixed typo in source where vtg source was plotting as gga HeadingTS: 1) added code to plot 2% threshold on mag error plot QRev: 1) fixed issue with sorting messages, 2) removed computation and use of self.invalid_bt ans self.invalid_gps, 3) fixed slow creatation of depth tab ShipTrack: 1) removed invalid_gps and invalid_bt from parameters and use of the parameter in the code, 2) fixed issues with plotting edge shiptracks, 3) modified subsection to handle nan, 4) added subsection_2d method

+++ Classes/GPSData.py
@@ -217,11 +217,11 @@ class GPSData(object):
         self.raw_gga_lon_deg[np.isnan(self.raw_gga_lat_deg)] = np.nan
         self.raw_gga_altitude_m = raw_gga_alt
         self.raw_gga_altitude_m[np.isnan(self.raw_gga_lat_deg)] = np.nan
-        self.raw_gga_differential = raw_gga_diff
+        self.raw_gga_differential = raw_gga_diff.astype('float')
         self.raw_gga_differential[np.isnan(self.raw_gga_lat_deg)] = np.nan
-        self.raw_gga_hdop = raw_gga_hdop
+        self.raw_gga_hdop = raw_gga_hdop.astype('float')
         self.raw_gga_hdop[np.isnan(self.raw_gga_lat_deg)] = np.nan
-        self.raw_gga_num_sats = raw_gga_num_sats
+        self.raw_gga_num_sats = raw_gga_num_sats.astype('float')
         self.raw_gga_num_sats[np.isnan(self.raw_gga_lat_deg)] = np.nan
         self.raw_gga_serial_time[np.isnan(self.raw_gga_lat_deg)] = np.nan
